+++ RELEASE_NOTES.md
@@ -1,13 +1,14 @@
-# Release 0.5.43
+# Release 0.5.44
 
 **Date:** 2026-04-29
 
 ## What's New
 
+- Merge feature 'gitflow-auto-version-fallback' into develop
+- Squash(feature): makefile-wizard-logic-fix
 - Squash(feature): tap-publish-validation-automation
-- Merge feature 'fix-sha' into develop
 
 ## Bug Fixes
 
-- Fix-sha
+- Fallback to next patch release version